requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
cachetools==7.0.1
    # via -r requirements.in
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.5
    # via pyjwt
fastapi==0.129.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pyjwt[crypto]==2.11.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via -r requirements.in
